Kill RotateCommand
